Use your phone to control a robot
This is a program used to control a robot through UDP, features include
a settings menu that allows the user to enter the robot's IP Address and
command socket. MJPEG video will be added and currently trying to add a
compass that shows the robot's heading.  Controller background is a
1280x720 PNG image that is not set up for dpi yet and only one
resolution image exist.